Swap landing and prelanding path
