--- src/lib/common/enums/path.ts
@@ -1,7 +1,0 @@
-export enum Path {
-	PRELANDING = '/prelanding',
-	LANDING = '/',
-	ABOUT = '/about',
-	CONTACT = '/contact',
-	BLOG = '/blog'
-}
